Replace swc with esbuild
