+++ pinefile.js
@@ -1,0 +1,4 @@
+import { run } from '@pinefile/pine';
+
+export default () => console.log('default');
+export const test = () => console.log('foo');
