+++ src/experiments/index.ts
@@ -1,0 +1,4 @@
+import { register as registerGlobalFile } from './global-file';
+
+export const register = (pine: any) =>
+  [registerGlobalFile].reduce((prev, cur) => cur(prev), pine);
